Implement a basic DataManager & general updates
-Create and Partially Implemented IDataManager
-Added necessary data to create view
-Added scheduler to nav bar
-Added SQL Procedure for Creating Scheduled Events
-Changed SQL Ids to BigInt

+++ ScheduleNetData/dbo/Tables/RequestedDates.sql
@@ -1,6 +1,6 @@
 ﻿CREATE TABLE [dbo].[RequestedDates]
 (
-	[Id] INT NOT NULL PRIMARY KEY, 
-    [ScheduledEventId] INT NOT NULL REFERENCES ScheduledEvents(Id) , 
+	[Id] BIGINT NOT NULL PRIMARY KEY, 
+    [ScheduledEventId] BIGINT NOT NULL REFERENCES ScheduledEvents(Id) , 
     [RequestedDate] DATETIME NOT NULL
 )
